// src/forgetrustee/frontend/src/types/generated/titleDescriptionRestDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface TitleDescriptionRestDto {
  /**
     * @minLength 0
     * @maxLength 100
     */
  title?: string;
  /**
     * @minLength 0
     * @maxLength 2000
     */
  description?: string;
}